+++ apps/docs/feature-flags/index.ts
@@ -1,2 +1,2 @@
 export type { FeatureFlag } from './featureFlags'
-export { featureFlags, defaultFlagConfig } from './featureFlags'
+export { featureFlags } from './featureFlags'
